refactor: nettoyage ancien composant
