+++ shared/constants/organismes.ts
@@ -1,0 +1,6 @@
+export const natureOrganismeDeFormationLabel = {
+  responsable: "Responsable",
+  formateur: "Formateur",
+  responsable_formateur: "Responsable et formateur",
+  inconnue: "Inconnue",
+};
